Remove circular dependancies
Turn was a global variable being exported, into the ui for declaring a
winner but this created a circular dependency. Discovered that it is
only been exported the way it is first declared in the events, and that
it would not carry its increments with it to the next page.